requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.0